rename PriceUtils, adjust testing
